+++ psyneulink/core/components/functions/__init__.py
@@ -1,5 +1,8 @@
 from . import function
+from . import userdefinedfunction
 
 from .function import *
+from .userdefinedfunction import *
 
 __all__ = list(function.__all__)
+__all__.extend(userdefinedfunction.__all__)
