+++ psyneulink/core/components/projections/modulatory/learningprojection.py
@@ -166,7 +166,8 @@ import numpy as np
 import typecheck as tc
 
 from psyneulink.core.components.component import parameter_keywords
-from psyneulink.core.components.functions.function import BackPropagation, Linear, LinearCombination, is_function_type
+from psyneulink.core.components.functions.function import BackPropagation, Linear, is_function_type
+from psyneulink.core.components.functions.combinationfunctions import LinearCombination
 from psyneulink.core.components.mechanisms.adaptive.learning.learningmechanism import ERROR_SIGNAL, LearningMechanism
 from psyneulink.core.components.mechanisms.processing.objectivemechanism import ObjectiveMechanism
 from psyneulink.core.components.projections.modulatory.modulatoryprojection import ModulatoryProjection_Base
