+++ psyneulink/library/components/mechanisms/adaptive/control/evc/evccontrolmechanism.py
@@ -379,7 +379,7 @@ import numpy as np
 import typecheck as tc
 
 from psyneulink.core.components.component import function_type
-from psyneulink.core.components.functions.function import LinearCombination
+from psyneulink.core.components.functions.combinationfunctions import LinearCombination
 from psyneulink.core.components.functions.function import ModulationParam, _is_modulation_param
 from psyneulink.core.components.mechanisms.adaptive.control.controlmechanism import ControlMechanism
 from psyneulink.core.components.mechanisms.mechanism import Mechanism, MechanismList
