+++ tests/mechanisms/test_processing_mechanism.py
@@ -1,7 +1,10 @@
 import numpy as np
 import pytest
 
-from psyneulink.core.components.functions.function import AGTUtilityIntegrator, AccumulatorIntegrator, AdaptiveIntegrator, BackPropagation, BogaczEtAl, CombineMeans, ConstantIntegrator, Distance, DriftDiffusionIntegrator, Exponential, ExponentialDist, FHNIntegrator, FunctionError, GammaDist, Hebbian, Integrator, Linear, LinearCombination, LinearMatrix, Logistic, NavarroAndFuss, NormalDist, OrnsteinUhlenbeckIntegrator, Reduce, Reinforcement, SimpleIntegrator, SoftMax, Stability, TDLearning, UniformDist, UniformToNormalDist, WaldDist
+from psyneulink.core.components.functions.function import AGTUtilityIntegrator, AccumulatorIntegrator, AdaptiveIntegrator, BackPropagation, BogaczEtAl, \
+    ConstantIntegrator, Distance, DriftDiffusionIntegrator, Exponential, ExponentialDist, FHNIntegrator, FunctionError, GammaDist, Hebbian, Integrator, Linear, \
+    LinearMatrix, Logistic, NavarroAndFuss, NormalDist, OrnsteinUhlenbeckIntegrator, Reinforcement, SimpleIntegrator, SoftMax, Stability, TDLearning, UniformDist, UniformToNormalDist, WaldDist
+from psyneulink.core.components.functions.combinationfunctions import Reduce, LinearCombination, CombineMeans
 from psyneulink.core.components.mechanisms.processing.processingmechanism import ProcessingMechanism
 
 class TestProcessingMechanismFunctions:
